+++ src/videocoin_proving_system/prover/prover.h
@@ -5,13 +5,11 @@
 #ifndef PROVER_H
 #define PROVER_H
 
-#include <common/utility.h>
-
 void initialize_prover();
 
 double generate_ssim_proof(const char *pk_fn,
-                           const unsigned char *src1, size_t src1_len,
-                           const unsigned char *src2, size_t src2_len,
+                           const unsigned char *src1, unsigned long src1_len,
+                           const unsigned char *src2, unsigned long src2_len,
                            const char *output_fn, const char *proof_fn,
                            const char *json_fn = nullptr);
